feature: New feature added: network latency for profile - 'slow'

+++ payment/src/main/kotlin/com/tkluza/smartcity/payment/business/external/reservation/config/ReservationExternalProperties.kt
@@ -5,5 +5,5 @@ import org.springframework.boot.context.properties.ConfigurationProperties
 @ConfigurationProperties("payment.http-client")
 data class ReservationExternalProperties(
     var smartMobilityApplicationUrl: String = "",
-    var networkLatencyMillis: Int = 0
+    var networkLatencyMillis: Long = 0
 )
